CI/CD設定

+++ db/migrate/20241231085335_remove_uid_from_users.rb
@@ -1,5 +1,7 @@
 class RemoveUidFromUsers < ActiveRecord::Migration[8.0]
   def change
-    remove_column :users, :uid, :string
+    if column_exists?(:users, :uid)
+      remove_column :users, :uid, :string
+    end
   end
 end
